Revert "API flagging macro until/starting_at for private and public"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6fa44cb3-33cc-4aef-bb9b-6b4454b3b637). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 363981334
Change-Id: I4e364c08c17c88642ab71a84d426f3297b57d6f3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com