Revert "API flagging macro until/starting_at for private and public"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/72078b1e-4ec8-47c9-90b9-c0c4aa7dd242). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 363980708
Change-Id: I6eef9bc9fb819d5d43599b24fb03553c5ee77ebf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com