Revert "API flagging macro until/starting_at for private and public"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f08a1ac5-7526-4473-ad32-b55565db2b49). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 363980804
Change-Id: I838cb34d9995879f2fb8d5f9cac6e327fb78afd8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com