Revert "API flagging macro until/starting_at for private and public"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/302a648c-62ec-45d0-9563-240ca926abc3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 363980889
Change-Id: Ifc00acb1afe9d33273373feba14bbb255ab3ae6b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com