Revert "API flagging macro until/starting_at for private and public"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6c50372d-2ec7-4acc-ab74-e248cfde7c6a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 363980909
Change-Id: I52e34e9d9c1e96ab9c848ea0c50c3bb67caf03c7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com