Revert "API flagging macro until/starting_at for private and public"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/15b3e219-fcc2-4328-b2f4-34ea9eb1aab8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 363980866
Change-Id: I5fd2917cdaa1162796b9a7d6f912095aeea2c259
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com