Define nsfs file type.
This file type was originally defined for goldfish but this is not
goldfish specific. We want to use this type in car_generic system
sepolicy as well, so we move this to system sepolicy.

This type is used to guard access to /proc/[pid]/ns.

Flag: EXEMPT selinux policy change
Test: Presubmit
Bug: 364652736
Change-Id: Ie03531372fc96a05486ff9d9cdfce79d2b0992b9