Revert "Add sepolicy for biometric fingerprint virtual hal"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/50c95536-897e-41ff-b630-ce3d1970967f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 364640436
Change-Id: I877b02923d0cbe1655d75356be6573dfd5c7ee66
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com