Revert "Add policy for 'debug.tracing.desktop_mode_visible_tasks..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/026ab552-7300-40b8-92b9-61f625e36550). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic49cc1765e223124947f327d18ef1fb3ef24cb85
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com